requirements.txt
Flask
Flask-MySQL
flask-restful
python-dotenv

